+++ doc/user_manual/snippets/test_algo.py
@@ -201,7 +201,7 @@ def test_compute_wall_distance():
   dist_tree = maia.io.file_to_dist_tree(mesh_dir/'U_ATB_45.yaml', MPI.COMM_WORLD)
   part_tree = maia.factory.partition_dist_tree(dist_tree, MPI.COMM_WORLD)
 
-  maia.algo.part.compute_wall_distance(part_tree, MPI.COMM_WORLD)
+  maia.algo.compute_wall_distance(part_tree, MPI.COMM_WORLD)
   assert maia.pytree.get_node_from_name(part_tree, "WallDistance") is not None
   #compute_wall_distance@end
 
@@ -212,7 +212,7 @@ def test_compute_iso_surface():
   from   maia.utils.test_utils import mesh_dir
   dist_tree = maia.io.file_to_dist_tree(mesh_dir/'U_ATB_45.yaml', MPI.COMM_WORLD)
   part_tree = maia.factory.partition_dist_tree(dist_tree, MPI.COMM_WORLD, preserve_orientation=True)
-  maia.algo.part.compute_wall_distance(part_tree, MPI.COMM_WORLD, point_cloud='Vertex')
+  maia.algo.compute_wall_distance(part_tree, MPI.COMM_WORLD, point_cloud='Vertex')
 
   part_tree_iso = maia.algo.part.iso_surface(part_tree, "WallDistance/TurbulentDistance", iso_val=0.25,\
       containers_name=['WallDistance'], comm=MPI.COMM_WORLD)
@@ -264,7 +264,7 @@ def test_extract_from_zsr():
   dist_tree = maia.io.file_to_dist_tree(mesh_dir/'U_ATB_45.yaml', MPI.COMM_WORLD)
   part_tree = maia.factory.partition_dist_tree(dist_tree, MPI.COMM_WORLD)
 
-  maia.algo.part.compute_wall_distance(part_tree, MPI.COMM_WORLD, point_cloud='Vertex')
+  maia.algo.compute_wall_distance(part_tree, MPI.COMM_WORLD, point_cloud='Vertex')
 
   # Create a ZoneSubRegion on procs for extracting odd cells
   for part_zone in PT.get_all_Zone_t(part_tree):
@@ -288,7 +288,7 @@ def test_extract_from_bc_name():
   dist_tree = maia.io.file_to_dist_tree(mesh_dir/'U_ATB_45.yaml', MPI.COMM_WORLD)
   part_tree = maia.factory.partition_dist_tree(dist_tree, MPI.COMM_WORLD)
 
-  maia.algo.part.compute_wall_distance(part_tree, MPI.COMM_WORLD, point_cloud='Vertex')
+  maia.algo.compute_wall_distance(part_tree, MPI.COMM_WORLD, point_cloud='Vertex')
 
   extracted_bc = maia.algo.part.extract_part_from_bc_name(part_tree, \
                  'wall', MPI.COMM_WORLD, containers_name=["WallDistance"])
@@ -305,7 +305,7 @@ def test_extract_from_family():
   dist_tree = maia.io.file_to_dist_tree(mesh_dir/'U_ATB_45.yaml', MPI.COMM_WORLD)
   part_tree = maia.factory.partition_dist_tree(dist_tree, MPI.COMM_WORLD)
 
-  maia.algo.part.compute_wall_distance(part_tree, MPI.COMM_WORLD, point_cloud='Vertex')
+  maia.algo.compute_wall_distance(part_tree, MPI.COMM_WORLD, point_cloud='Vertex')
 
   extracted_bc = maia.algo.part.extract_part_from_family(part_tree, \
                  'WALL', MPI.COMM_WORLD, containers_name=["WallDistance"])